+++ src/Value/Position/Lambert72Point.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Position;
+
+use DigipolisGent\Value\ValueAbstract;
+
+/**
+ * Lambert 72 representation of an address position point.
+ */
+final class Lambert72Point extends AbstractPoint
+{
+
+}