+++ src/Value/Position/PointInterface.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Position;
+
+/**
+ * Interface for address position point.
+ */
+interface PointInterface
+{
+    /**
+     * Get the x position.
+     *
+     * @return float
+     */
+    public function xPosition();
+
+    /**
+     * Get the y position.
+     *
+     * @return float
+     */
+    public function yPosition();
+}