+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All Notable changes to `digipolisgent/flanders-basicregisters` package.
 
 * Added the Locality value object.
 * Added the StreetName value object.
+* Added the address point values.
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/master...develop